requirements.txt
numpy
pandas
matplotlib
scikit-learn
streamlit
pytest
scipy
seaborn
plotly
requests
openai
langchain
langgraph
chromadb
pydantic
httpx
beautifulsoup4
structlog
python-dotenv
networkx
python-dateutil